Refactored Farajaland repo and statistics import
