+++ db-populate.sh
@@ -10,7 +10,7 @@
 # graphic logo are (registered/a) trademark(s) of Plan International.
 
 print_usage_and_exit () {
-    echo 'Usage: ./db-populate.sh --password_for_users=XXX --environment=development|production --alpha3_country_code=XXX'
+    echo 'Usage: ./db-populate.sh --password_for_users=XXX --environment=development|production --alpha3_country_code=XXX --path_to_admin_structure_csv=XXX --path_to_crvs_offices_csv=XXX --path_to_health_facilities_csv=XXX --path_to_statistics_csv=XXX --path_to_employees_csv=XXX'
     echo "  --password_for_users will be the password and security question answer for all users used in development."
     echo "    The --password_for_users property is ignored if populating production users in bulk."
     echo "    In production, first time use passwords will be dynamically generated by this script."
@@ -34,6 +34,26 @@ for i in "$@"; do
       alpha3_country_code="${i#*=}"
       shift
       ;;
+    --path_to_admin_structure_csv=*)
+      path_to_admin_structure_csv="${i#*=}"
+      shift
+      ;;
+    --path_to_statistics_csv=*)
+      path_to_statistics_csv="${i#*=}"
+      shift
+      ;;
+    --path_to_crvs_offices_csv=*)
+      path_to_crvs_offices_csv="${i#*=}"
+      shift
+      ;;
+    --path_to_health_facilities_csv=*)
+      path_to_health_facilities_csv="${i#*=}"
+      shift
+      ;;
+    --path_to_employees_csv=*)
+      path_to_employees_csv="${i#*=}"
+      shift
+      ;;
     -*|--*)
       echo "Unknown option $i"
       exit 1
@@ -43,7 +63,15 @@ for i in "$@"; do
   esac
 done
 
-
+# Retrieve 2-step verification to continue
+#-----------------------------------------
+ask_yes_or_no() {
+    read -p "$1 ([y]es or [N]o): "
+    case $(echo $REPLY | tr '[A-Z]' '[a-z]') in
+        y|yes) echo "yes" ;;
+        *)     echo "no" ;;
+    esac
+}
 
 if [ -z $environment ] || { [ $environment != 'development' ] && [ $environment != 'production' ] ;} ; then
     echo "MISSING OR INCORRECT PARAMETER: --environment"
@@ -57,6 +85,34 @@ elif [ -z $alpha3_country_code ]; then
     echo "MISSING OR INCORRECT PARAMETER: --alpha3_country_code"
     echo
     print_usage_and_exit
+elif [ -z $path_to_admin_structure_csv ]; then
+    echo "MISSING OR INCORRECT PARAMETER: --path_to_admin_structure_csv"
+    echo
+    print_usage_and_exit
+elif [ -z $path_to_crvs_offices_csv ]; then
+    echo "MISSING OR INCORRECT PARAMETER: --path_to_crvs_offices_csv"
+    echo
+    print_usage_and_exit
+elif [ -z $path_to_health_facilities_csv ]; then
+    echo "MISSING OR INCORRECT PARAMETER: --path_to_health_facilities_csv"
+    echo
+    print_usage_and_exit
+elif [ -z $path_to_employees_csv ]; then
+    echo "MISSING OR INCORRECT PARAMETER: --path_to_employees_csv"
+    echo
+    print_usage_and_exit
+elif [ -z $path_to_statistics_csv ]; then
+    echo "MISSING OR INCORRECT PARAMETER: --path_to_statistics_csv"
+    echo
+    echo "STATISTICS CSV HAS NOT BEEN SUPPLIED.  Do you wish to proceed without statistics?"
+    echo
+    echo "Please note that the Performance Dashboard will be unable to calculate Registration Completeness Rates without statistics."
+    echo
+    if [[ "no" == $(ask_yes_or_no "If you are ready to continue, type: yes.  If you dont know, type: no to exit.") ]]
+    then
+        echo -e "\n\nExiting."
+        exit 0
+    fi
 elif [ -z $password_for_users ] && [ $environment == 'production' ] ; then
     echo "PROCEEDING TO POPULATE DATABASE WITH BULK USERS IN PRODUCTION."
     echo
@@ -83,11 +139,8 @@ docker run --rm --network=$NETWORK appropriate/curl curl -X POST 'http://influxd
 
 
 # Populate new application data
-ts-node -r tsconfig-paths/register src/scripts/validate-source-files.ts
-ts-node -r tsconfig-paths/register src/features/administrative/scripts/prepare-locations.ts
-ts-node -r tsconfig-paths/register src/features/administrative/scripts/assign-admin-structure-to-locations.ts
-ts-node -r tsconfig-paths/register src/features/facilities/scripts/prepare-source-facilities.ts
-ts-node -r tsconfig-paths/register src/features/facilities/scripts/assign-facilities-to-locations.ts
-ts-node -r tsconfig-paths/register src/features/administrative/scripts/add-statistical-data.ts
-ts-node -r tsconfig-paths/register src/features/employees/scripts/prepare-source-employees.ts -- $environment
-ts-node -r tsconfig-paths/register src/features/employees/scripts/assign-employees-to-practitioners.ts -- $password_for_users $environment $alpha3_country_code
+ts-node -r tsconfig-paths/register src/scripts/validate-source-files.ts -- $path_to_admin_structure_csv $path_to_crvs_offices_csv $path_to_health_facilities_csv $path_to_employees_csv $path_to_statistics_csv
+ts-node -r tsconfig-paths/register src/features/administrative/scripts/import-admin-structure.ts -- $path_to_admin_structure_csv
+ts-node -r tsconfig-paths/register src/features/facilities/scripts/import-offices-and-health-facilities.ts -- $path_to_crvs_offices_csv $path_to_health_facilities_csv
+ts-node -r tsconfig-paths/register src/features/administrative/scripts/import-statistics.ts -- $path_to_statistics_csv
+ts-node -r tsconfig-paths/register src/features/employees/scripts/import-employees.ts -- $path_to_employees_csv $password_for_users $environment $alpha3_country_code
