--- src/features/administrative/source/districts.csv
@@ -1,17 +1,0 @@
-statisticalID,name,partOf,code,physicalType
-oEBf29y8JP8,Ibombo,Location/AWn3s2RqgAN,ADMIN_STRUCTURE,Jurisdiction
-HPGiE9Jjh2r,Isamba,Location/AWn3s2RqgAN,ADMIN_STRUCTURE,Jurisdiction
-BxrIbNW7f3K,Itambo,Location/AWn3s2RqgAN,ADMIN_STRUCTURE,Jurisdiction
-SQT8xjbvWwf,Ezhi,Location/AWn3s2RqgAN,ADMIN_STRUCTURE,Jurisdiction
-ntoX1PkiWri,Ilanga,Location/KozcEjeTyuD,ADMIN_STRUCTURE,Jurisdiction
-SvFNQpplnch,Irundu,Location/KozcEjeTyuD,ADMIN_STRUCTURE,Jurisdiction
-NLjvK1QsrN3,Zobwe,Location/KozcEjeTyuD,ADMIN_STRUCTURE,Jurisdiction
-hywxVKv48Xp,Afue,Location/KozcEjeTyuD,ADMIN_STRUCTURE,Jurisdiction
-QTtxiWj8ONP,Embe,Location/B1u1bVtIA92,ADMIN_STRUCTURE,Jurisdiction
-Oapn6R4rt3D,Ienge,Location/B1u1bVtIA92,ADMIN_STRUCTURE,Jurisdiction
-idR0pJWLqcR,Funabuli,Location/B1u1bVtIA92,ADMIN_STRUCTURE,Jurisdiction
-jCJxK8sFH40,Pili,Location/B1u1bVtIA92,ADMIN_STRUCTURE,Jurisdiction
-ydyJb1RAy4U,Ama,Location/dbTLdTi7s8F,ADMIN_STRUCTURE,Jurisdiction
-nO68NTtcdw2,Nsali,Location/dbTLdTi7s8F,ADMIN_STRUCTURE,Jurisdiction
-FuvOOwvBw0p,Soka,Location/dbTLdTi7s8F,ADMIN_STRUCTURE,Jurisdiction
-R22tRIhdm5Y,Chibiya,Location/dbTLdTi7s8F,ADMIN_STRUCTURE,Jurisdiction
