+++ src/features/administrative/source/farajaland.csv
@@ -1,4 +1,4 @@
-admin2Name_en,admin2Name_ar,admin2Pcode,admin1Name_en,admin1Name_ar,admin1Pcode,admin0Name,admin0Name_ar,admin0Pcode
+admin2Name_en,admin2Name_alias,admin2Pcode,admin1Name_en,admin1Name_alias,admin1Pcode,admin0Name,admin0Name_alias,admin0Pcode
 Ibombo,الحلة,oEBf29y8JP8,Central,بابل,AWn3s2RqgAN,Farajaland,العراق,FAR
 Ilanga,الدور,ntoX1PkiWri,Sulaka,صلاح الدين,KozcEjeTyuD,Farajaland,العراق,FAR
 Embe,الكاظمية,QTtxiWj8ONP,Pualula,بغداد,B1u1bVtIA92,Farajaland,العراق,FAR
