+++ src/features/administrative/source/jamaica.csv
@@ -1,0 +1,15 @@
+admin0Name_en,admin0Pcode,admin1Name_en,admin1Pcode
+Jamaica,JM,Clarendon,JM01
+Jamaica,JM,Hanover,JM02
+Jamaica,JM,Manchester,JM04
+Jamaica,JM,Portland,JM07
+Jamaica,JM,Saint Andrew,JM08
+Jamaica,JM,Saint Ann,JM09
+Jamaica,JM,Saint Catherine,JM10
+Jamaica,JM,Saint Elizabeth,JM11
+Jamaica,JM,Saint James,JM12
+Jamaica,JM,Saint Mary,JM13
+Jamaica,JM,Saint Thomas,JM14
+Jamaica,JM,Trelawny,JM15
+Jamaica,JM,Westmoreland,JM16
+Jamaica,JM,Kingston,JM17
