--- src/features/administrative/source/states.csv
@@ -1,5 +1,0 @@
-statisticalID,name,partOf,code,physicalType
-AWn3s2RqgAN,Central,Location/0,ADMIN_STRUCTURE,Jurisdiction
-KozcEjeTyuD,Sulaka,Location/0,ADMIN_STRUCTURE,Jurisdiction
-B1u1bVtIA92,Pualula,Location/0,ADMIN_STRUCTURE,Jurisdiction
-dbTLdTi7s8F,Chuminga,Location/0,ADMIN_STRUCTURE,Jurisdiction
