--- src/features/administrative/source/statistics.csv
@@ -1,17 +1,0 @@
-statisticalID,name,male_population_2007,female_population_2007,population_2007,crude_birth_rate_2007,male_population_2008,female_population_2008,population_2008,crude_birth_rate_2008,male_population_2009,female_population_2009,population_2009,crude_birth_rate_2009,male_population_2010,female_population_2010,population_2010,crude_birth_rate_2010,male_population_2011,female_population_2011,population_2011,crude_birth_rate_2011,male_population_2012,female_population_2012,population_2012,crude_birth_rate_2012,male_population_2013,female_population_2013,population_2013,crude_birth_rate_2013,male_population_2014,female_population_2014,population_2014,crude_birth_rate_2014,male_population_2015,female_population_2015,population_2015,crude_birth_rate_2015,male_population_2016,female_population_2016,population_2016,crude_birth_rate_2016,male_population_2017,female_population_2017,population_2017,crude_birth_rate_2017,male_population_2018,female_population_2018,population_2018,crude_birth_rate_2018,male_population_2019,female_population_2019,population_2019,crude_birth_rate_2019,male_population_2020,female_population_2020,population_2020,crude_birth_rate_2020,male_population_2021,female_population_2021,population_2021,crude_birth_rate_2021
-oEBf29y8JP8,Ibombo,40971,39344,80323,18.2,42668,41823,84511,21,45008,43843,88871,21.8,45150,43903,89073,16.2,45513,43884,89409,12.6,47965,46643,94631,20.6,25192,24542,49734,17.5,25659,25277,50936,18.9,32924,32808,65731,16.5,34031,34013,68044,12.6,33472,32968,66440,16.7,32876,28563,61440,13.8,36647,27595,64242,11.1,39380,32952,72332,11.7,46278,27705,73983,10.5
-HPGiE9Jjh2r,Isamba,69782,64619,134371,19.5,67498,64096,131590,18.6,68665,64425,133069,14.7,69810,65957,135754,18.4,70091,66079,136153,27.2,71602,66857,138426,19.4,65834,63461,129296,17.7,53334,54067,107402,19.6,190512,193439,383952,16.5,192934,195623,388557,17.2,179651,179636,359287,19.5,200566,199729,400295,17.9,235281,197258,432538,14.6,249443,211685,461128,12.3,245007,184189,429196,13.9
-BxrIbNW7f3K,Itambo,60206,53830,113974,20.7,59009,52811,111762,23.1,58243,52139,110310,17.3,58014,52207,110151,17.6,60815,54380,115111,19.8,60648,53659,114206,18.2,51151,47880,99033,21.2,41440,39577,81016,21.7,68791,66503,135293,22.7,73056,70325,143382,22,73629,70894,144524,12.4,70410,64749,135159,12.5,75221,56581,131801,12.2,60738,60791,121529,13.5,70895,61882,132777,15.2
-SQT8xjbvWwf,Ezhi-tezhi,43316,42961,86309,21,38253,37439,75709,21.2,53334,51972,105329,18.4,45577,45721,91349,17.8,46149,45840,92033,18.7,44235,43929,88208,21.3,20360,19571,39932,19.2,20241,19974,40215,20.6,29049,28792,57841,17.6,29248,28758,58006,18.6,28260,27756,56016,17.2,25545,27201,52746,18.8,25659,26793,52452,19.4,23434,24550,47984,17.2,24982,25295,50278,16.3
-ntoX1PkiWri,Ilanga,39262,37358,76621,19.3,40044,38649,78701,20.2,41564,39172,80727,19,41392,39070,80453,18.7,41508,39270,80770,15.5,42314,39919,82223,10.9,39941,39287,79228,15,38917,38045,76962,15.9,39040,39384,78424,18.8,39520,39049,78569,17.7,40754,40597,81351,18.3,46524,41900,88424,17.3,54068,46751,100819,19.3,64557,50394,114952,21.9,56918,41369,98287,24
-SvFNQpplnch,Irundu,41723,38091,79785,18.4,43345,38834,82137,21.2,44064,40065,84088,18.8,43897,40288,84150,20,44647,41206,85821,19.7,45958,41839,87749,20.2,30223,29491,59714,17.8,30029,30017,60047,20.1,34387,34904,69291,21.8,34115,34323,68439,27.8,32550,32630,65180,23.3,30169,35643,65813,22.7,32508,29253,61761,19.6,33079,30747,63826,16.9,36354,27886,64240,19.4
-NLjvK1QsrN3,Zobwe,68976,65286,134256,16.9,70717,66503,137205,22.8,72297,68143,140423,17.6,72625,68558,141169,18.5,73523,69672,143185,17.8,74485,71308,145801,18.3,74706,72935,147641,18,75411,74447,149858,15.6,74552,73697,148248,13.5,72687,72264,144951,18.4,71591,71303,142894,16.9,62110,84182,146292,13.7,69080,90774,159854,11.8,58136,102480,160616,11.6,60655,112231,172886,11.4
-hywxVKv48Xp,Afue,52065,48359,100403,16.2,52059,48619,100661,12,52813,49667,102465,15.8,53526,50534,104050,18.4,54076,51129,105195,15.3,54833,52506,107345,15.3,47968,45428,93397,14.9,47793,45161,92953,13.5,51881,50619,102500,14.9,52210,50859,103069,12.8,47997,47388,95386,16.7,46462,38863,85325,17.1,39956,33855,73811,18.5,32659,34539,67199,22.1,38938,40308,79246,23.2
-QTtxiWj8ONP,Embe,44191,44208,88438,22.1,44862,44626,89520,18.4,45647,45329,91014,18,46251,45707,91996,18.9,47156,46207,93393,18.3,48513,47973,96529,19.3,19871,20162,40032,18.1,20207,20830,41037,22.6,17181,17473,34654,19.4,17381,18022,35403,17.8,17447,17663,35110,17.6,16251,16120,32371,14.3,13111,16234,29345,17,10795,16609,27404,19,9870,19882,29752,17.7
-Oapn6R4rt3D,Ienge,45749,43132,88875,19.5,46372,43244,89601,20,47251,43696,90919,17.6,47497,44443,91922,16.8,50903,47897,98785,17.1,48164,45299,93449,13.2,56005,54088,110093,14.9,55970,54585,110555,14.7,50231,49359,99591,19.2,51345,50087,101432,15.6,49516,48123,97640,19.8,41502,52584,94086,22.9,46480,47181,93662,20.5,46726,56442,103167,18.4,43100,55627,98727,20.7
-idR0pJWLqcR,Funabuli,69454,65596,135042,20.1,67792,63091,130857,20.8,71984,67077,139029,19.5,72715,68038,140724,18.1,74718,68937,143602,16.1,76357,70161,146455,17.1,91749,89408,181157,18,93096,92563,185659,18.1,86496,85278,171774,18.4,87701,87432,175134,19,84150,83384,167534,20.7,99309,79972,179282,20.1,98410,92248,190658,18.4,98251,96108,194359,20,82404,82479,164883,19.9
-jCJxK8sFH40,Pili,40322,37127,77427,17.9,38697,35747,74425,18.3,39905,37213,77101,18.6,40098,37272,77349,20.5,41417,38862,80265,21.9,41089,38521,79595,22.4,24835,23996,48831,17.6,25468,24388,49855,22.9,23145,22402,45546,20.4,18959,18324,37283,25.5,21672,21166,42838,20.9,20045,17459,37504,22.7,22825,16121,38946,18.2,25150,13315,38465,19,25727,11500,37227,20.2
-ydyJb1RAy4U,Ama,53131,49864,102983,17.8,54102,50457,104541,18.8,55868,52772,108629,16.4,56458,53491,109942,16.6,57131,54567,111701,20.1,58741,55824,114561,20.6,58821,57242,116064,19.8,60881,57907,118787,19.8,59867,57145,117011,16,60854,57961,118815,18.4,61186,58969,120156,20,59417,52603,112019,22.2,49048,53876,102925,21.7,48659,48973,97632,17.9,44180,40980,85159,15.1
-nO68NTtcdw2,Nsali,43774,40176,83924,18.7,43864,41010,84860,24.6,45064,41631,86667,19.4,45831,41776,87565,19.3,47708,43837,91507,17.1,48050,43752,91753,23.2,35210,33492,68703,21.6,36236,34510,70745,20.5,30377,29046,59423,21.4,30716,29371,60087,25.9,31012,29510,60523,17.7,31506,32908,64414,16.4,33228,33203,66431,17.5,35994,31242,67236,19,31786,29832,61617,17.6
-FuvOOwvBw0p,Soka,47198,44934,92133,26.3,48402,45723,94119,23.9,50237,47581,97811,21.6,50511,47479,97976,23.8,51644,48765,100398,23.2,52933,49975,102896,18.3,49704,48828,98532,17.9,51273,50723,101996,22.2,50680,49951,100631,20.7,51387,50356,101743,20.8,49360,47934,97295,20.2,53407,40728,94135,22.3,58185,46204,104390,18.1,52587,43979,96565,16.1,46290,51781,98071,17.8
-R22tRIhdm5Y,Chibiya,43692,40928,84609,23.4,44158,40481,84611,25.7,45432,42704,88123,22.9,46664,43172,89807,27.1,47827,43650,91431,22.4,49011,44397,93352,28.9,27986,26961,54946,20.8,28103,27529,55632,20,29904,28378,58281,18.3,27670,26223,53893,18.7,26429,25039,51470,21.2,27443,30018,57460,19.6,23001,35855,58856,21.1,18508,34629,53137,24.7,21629,37207,58836,24.1
