+++ src/features/facilities/source/health-facilities.csv
@@ -1,4 +1,4 @@
-statisticalID,name,partOf,code,physicalType
+adminPcode,name,partOf,code,physicalType
 pXhz0PLiYZX,Chamakubi Health Post,Location/oEBf29y8JP8,HEALTH_FACILITY
 di3U5u7F8Y3,Ibombo Rural Health Centre,Location/oEBf29y8JP8,HEALTH_FACILITY
 B5LpoYehUfI,Chikobo Rural Health Centre,Location/oEBf29y8JP8,HEALTH_FACILITY
