+++ src/features/generateRegistrationNumber/service.ts
