+++ src/features/languages/content/client/contentful-keys-to-migrate.json
